sử dụng container làm trung gian để lấy  dữ liệu từ store và truyền dữ liệu xuống products
